move BiscuitBuilder to its own file
